+++ src/index.ts
@@ -1,1 +1,4 @@
 export * from './components';
+
+// Import styles
+import './styles/index.scss';